requirements.txt
Flask==2.2.5
paho-mqtt==1.5.1
Werkzeug==2.3.7 